+++ klibs/KLExceptions.py
@@ -27,6 +27,21 @@ class TrialException(Exception):
     def __str__(self):
         return self.message
 
+class TerminateBlock(Exception):
+    """Exception that ends a block of trials early when raised.
+
+    Intended for studies using titration or other situations where block length is
+    based on participant performance and should end when a threshold is met.
+
+    .. note:: When TermiateBlock is raised the block will end immediately, meaning that
+              if raised within a trial any data from that trial will not be saved.
+              To ensure that data for the final trial is recorded, you can check for the
+              termination criteria and raise the exception within `self.trial_prep()`.
+
+    """
+    def __init__(self, msg=""):
+        self.message = msg
+
 class EyeTrackerError(Exception):
     """Raised when a problem relating to an :obj:`~klibs.KLEyeTracking.KLEyeTracker.EyeTracker`
     object or the misuse of eye event inspect/report types is encountered.